Add cylinder check. Add sphere check.

+++ include/Sphere.hpp
@@ -18,6 +18,6 @@ public:
 
 private:
     mwTPoint3d<double> center_;
-    float radius_;
+    double radius_;
 };
 }
